requirements.txt
astroid==2.4.0
attrs==19.3.0
aws-sam-translator==1.29.0
aws-xray-sdk==2.6.0
awscli==1.18.61
boto==2.49.0
boto3==1.13.1
botocore==1.16.1
certifi==2020.4.5.1
cffi==1.14.0
cfn-lint==0.40.0
chardet==3.0.4
colorama==0.4.3
coverage==5.2
crypto==1.4.1
cryptography==2.9.2
DateTime==4.3
decorator==4.4.2
docker==4.3.1
docutils==0.15.2
ecdsa==0.14.1
future==0.18.2
idna==2.9
importlib-metadata==2.0.0
isort==4.3.21
Jinja2==2.11.2
jmespath==0.9.5
jsondiff==1.2.0
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.4.0
moto==1.3.16
Naked==0.1.31
networkx==2.5
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==5.4.3
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
responses==0.12.0
rsa==3.4.2
s3transfer==0.3.3
shellescape==3.8.1
six==1.14.0
sshpubkeys==3.1.0
toml==0.10.0
urllib3==1.25.9
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.4.0
zope.interface==5.1.0
